simplified pre  startup checks execution: removed container need and dropped useless checks
