+++ src/Transport/Topology/Topologies/ForwardingTopology.cs
@@ -10,7 +10,7 @@ namespace NServiceBus.AzureServiceBus
         ITopologySectionManager topologySectionManager;
         ITransportPartsContainer container;
 
-        public ForwardingTopology() : this(new TransportPartsContainer()){ }
+        public ForwardingTopology() : this(new TransportPartsContainer()) { }
 
         internal ForwardingTopology(ITransportPartsContainer container)
         {
@@ -86,9 +86,6 @@ namespace NServiceBus.AzureServiceBus
 
             var validationStrategyType = (Type)settings.Get(WellKnownConfigurationKeys.Topology.Addressing.Validation.Strategy);
             container.Register(validationStrategyType);
-
-           container.Register(typeof(NamespacesConfigurationCheck));
-           container.Register(typeof(ManageRightsCheck));
         }
 
         public Func<ICreateQueues> GetQueueCreatorFactory()
@@ -112,10 +109,9 @@ namespace NServiceBus.AzureServiceBus
 
         public StartupCheckResult ApplyPreStartupChecks()
         {
-            var checks = container.ResolveAll<IStartupCheck>();
-            var compositeStartupChecker = new CompositeStartupCheck(checks);
+            var check = new ManageRightsCheck(this.container);
 
-            return compositeStartupChecker.Apply();
+            return check.Apply();
         }
 
         public IManageSubscriptions GetSubscriptionManager()
